one-way retro
basically character undo, in the absence of achronal effects, needs to undo all the character's moves

which means if the character undoes through a one-way, then that undo WOULD have been valid, so we allow it

but we also ban the CONVERSE case that WOULDN'T have been valid as a result! the one-ways are now kind of like time gates. isn't that cool?

I might later make a different kind of one-way that works as a ratchet. maybe that's what green ones do?